+++ src/app/components/conversation/conversation.component.ts
@@ -131,7 +131,6 @@ export class ConversationComponent implements OnInit, AfterViewInit, OnChanges {
 
   setTimeoutSound: NodeJS.Timer;
 
-  showSpinner = false;
 
   constructor(
     public el: ElementRef,
@@ -179,7 +178,6 @@ export class ConversationComponent implements OnInit, AfterViewInit, OnChanges {
   // }
 
   ngAfterViewInit() {
-    this.isShowSpinner();
     this.g.currentConversationComponent = this;
     if (this.g.newConversationStart === true) {
       this.onNewConversationComponentInit();
@@ -1591,16 +1589,4 @@ export class ConversationComponent implements OnInit, AfterViewInit, OnChanges {
     }
   }
   // ========= END:: TRIGGER FUNCTIONS ============//
-
-  public isShowSpinner() {
-    const that = this;
-    if (this.messages && this.messages.length === 0) {
-      that.showSpinner = true;
-      setTimeout(() => {
-        that.showSpinner = false;
-      }, 5000);
-    } else {
-      that.showSpinner = false;
-    }
-  }
 }